Update patch set 3
Patch Set 3:

Hi Vineet, Santosh,

This is the issue I was referring to on the architecture call. Please +1 and merge if acceptable. The basic test is to see that solutionId, userId (that deployed the model) and revisionId are in the nginx log configuration. This is very important to the Clio point release. 

Thanks,

Justin

Patch-set: 3